% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRescueRFunctions.R
\name{SigRescueAnalyze}
\alias{SigRescueAnalyze}
\title{Extracts posterior sampling from \code{SigRescueR()}.}
\usage{
SigRescueAnalyze(res, ci = 0.025)
}
\arguments{
\item{res}{Numeric value indicating the probability mass of the credible interval used to
summarize the posterior (e.g. 0.025 is 2.5\% credible interval). Default = 0.025.}

\item{MutationType}{A character vector specifying the ordered mutation context.}
}
\description{
This function extract the posterior samples from the Bayesian model and computes the
posterior mean for the exposure and background while providing some reconstruction
metrics and generates the exposure mutational catalogue.
}
