% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRescueRFunctions.R
\name{SigRescuePlot}
\alias{SigRescuePlot}
\title{Plot SBS results.}
\usage{
SigRescuePlot(
  res,
  ci = 0.025,
  output_path = getwd(),
  filename = "clean_res",
  dpi = 600
)
}
\arguments{
\item{res}{An object returned by \code{SigRescueR()} containing posterior estimates.}

\item{ci}{Numeric value indicating the probability mass of the credible interval used to
summarize the posterior (e.g. 0.025 is 2.5\% credible interval). Default = 0.025.}

\item{output_path}{Character string specifying the full path where the ouput \code{.pdf} file
will be saved. Default = current working directory.}

\item{filename}{Character string specifying the file name. Default = "clean_res".}

\item{dpi}{Numeric value indicating the resolution (dots per inch) for the saved plot.
Default = 600.}
}
\description{
This function is designed to generate a visualization of the results from \code{SigRescueR()}.
SBS-96 mutational signature plots for the original (top left), reconstructed (bottom left),
background (top right) and inferred exposure (bottom right). Alpha transparency is set to
reflect background (0.5) and exposure (1). The total number of mutations is indicated within
the plots. A table with Reconstruction Accuracy (measured by cosine similarity), along with
the background and exposure activity.
}
