% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRescueRFunctions.R
\name{SigRescueR}
\alias{SigRescueR}
\title{Execute Bayesian inference}
\usage{
SigRescueR(
  objects,
  script_path = system.file("scripts", "model_runner.R", package = "SigRescueR")
)
}
\arguments{
\item{objects}{A list of objects from the global environment that are temporary saved  and
used for Bayesian inference. The list must include \code{model}, \code{treatment}, \code{control}, \code{output_path},
and \code{filename}. \code{model} is an object that contains Stan model configuration, such as number
of warm-ups, iterations, and chains, generated by \code{SigRescueRModel()}. \code{treatment} is a data
frame representing the mutational catalogue for exposed samples. Each column corresponds to a
sample and must include a \code{MutationType} column (first column) specifying the mutation context
(e.g. \verb{A[C>A]A}), and all other values must be integers. \code{control} is a data frame representing
the mutational catalogue for control samples. Similar to treatment, and must include a
\code{MutationType} column (first column) matching the order of the treatment, and all values must be
normalized (e.g. relative frequencies). \code{output_path} is a character string specifying the full
path where the output \code{.rda} file will be saved. \code{filename} is a character string specifying the
file name.}

\item{script_path}{A character string specifying the R script to run the model. Default = model_runner.R}
}
\description{
This function creates temporary \code{.rds} files and R script used to run Bayesian inference
using \code{model_runner.R}. It prepares model data and launch a R process.
}
