% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRescueRFunctions.R
\name{SigRescueRExtract}
\alias{SigRescueRExtract}
\title{Extracts posterior sampling from \code{SigRescueR()}.}
\usage{
SigRescueRExtract(res, MutationType)
}
\arguments{
\item{res}{An object returned by \code{SigRescueR()} containing posterior estimates.}

\item{MutationType}{A character vector specifying the ordered mutation context.}
}
\description{
This function extract the posterior samples from the Bayesian model and computes the
posterior mean for the exposure and background while providing some reconstruction
metrics and generates the exposure mutational catalogue.
}
