% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRescueRFunctions.R
\name{SigRescueSetup}
\alias{SigRescueSetup}
\title{Create object that contains Stan model configuration for \code{SigRescueR()}}
\usage{
SigRescueSetup(model = "COM-Poisson", warmup = 1000, iter = 2500, chains = 4)
}
\arguments{
\item{model}{A character string specifying the Stan model to use. Default = COM-Poisson.}

\item{warmup}{An integer to specify the number of warm-up iterations before sampling
begins. Default = 1000.}

\item{iter}{An integer to specify the total number of iterations. The number of posterior
sampling is equal to the total number of iterations - the number of warm-up iterations.
Default = 2500. Thus, 1500 posterior sampling.}

\item{chains}{An integer specifying the number of Markov chains to run. Default = 4.}
}
\description{
This function creates an object that preloads the selected Stan model with Stan model
configurations. By preloading the Stan model, it does not need to be loaded internally
each time Bayesian inference is used.
}
