% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRescueRFunctions.R
\name{SigRescuePlot}
\alias{SigRescuePlot}
\title{Plot SBS results.}
\usage{
SigRescuePlot(clean, output_path = ".", filename = "clean_res", dpi = 600)
}
\arguments{
\item{clean}{Data frame representing the mutational catalogue. Each column corresponds
to a sample and must include a \code{MutationType} column (first column) specifying the
mutation context (e.g. \verb{A[C>A]A}).}

\item{output_path}{Character string specifying the full path where the ouput \code{.pdf} file
will be saved. Default = ".".}

\item{filename}{Character string specifying the file name. Default = "clean_res".}

\item{dpi}{Numeric value indicating the resolution (dots per inch) for the saved plot.
Default = 600.}
}
\description{
This function is designed to generate a visualization of the results from \code{SigRescueR()}
after data extraction (\code{SigRescueRExtract()}) using \code{samp.cleaned}, where the x-axis
represents the mutation context faceted by mutation type and y-axis represents percentage
of SBS mutation.
}
